+++ pkg/common/url.go
@@ -70,6 +70,24 @@ func BuildPodURL(externalUrl, urlType string, stub *types.StubWithRelated, stubC
 	return url
 }
 
+func BuildPodDeploymentURL(externalUrl, urlType string, deployment *types.Deployment, stubConfig *types.StubConfigV1) string {
+	parsedUrl, err := url.Parse(externalUrl)
+	if err != nil {
+		return ""
+	}
+
+	portPlaceholder := "<PORT>"
+	if len(stubConfig.Ports) == 1 {
+		portPlaceholder = fmt.Sprintf("%d", stubConfig.Ports[0])
+	}
+
+	if urlType == InvokeUrlTypeHost {
+		return fmt.Sprintf("%s://%s-latest-%s.%s", parsedUrl.Scheme, deployment.Subdomain, portPlaceholder, parsedUrl.Host)
+	}
+
+	return fmt.Sprintf("%s://%s/pod/%s/latest/%s", parsedUrl.Scheme, parsedUrl.Host, deployment.Name, portPlaceholder)
+}
+
 func BuildSandboxURL(externalUrl, urlType string, stub *types.StubWithRelated, containerId string, port int32) string {
 	parsedUrl, err := url.Parse(externalUrl)
 	if err != nil {
