+++ pkg/types/event.go
@@ -61,10 +61,11 @@ var (
 type CacheContentKind string
 
 const (
-	CacheContentKindClipV1     CacheContentKind = "clip_v1"
-	CacheContentKindClipV2     CacheContentKind = "clip_v2"
-	CacheContentKindVolume     CacheContentKind = "volume"
-	CacheContentKindCheckpoint CacheContentKind = "checkpoint"
+	CacheContentKindClipV1       CacheContentKind = "clip_v1"
+	CacheContentKindClipV2       CacheContentKind = "clip_v2"
+	CacheContentKindVolume       CacheContentKind = "volume"
+	CacheContentKindCheckpoint   CacheContentKind = "checkpoint"
+	CacheContentKindDiskSnapshot CacheContentKind = "disk_snapshot"
 )
 
 // Platform cache audit statuses for EventPlatformCacheSchema.Status.
@@ -91,6 +92,7 @@ type CacheRequiredContentItem struct {
 	ExpectedHash string           `json:"expected_hash,omitempty"`
 	ImageID      string           `json:"image_id,omitempty"`
 	Source       string           `json:"source,omitempty"`
+	SourceBucket string           `json:"source_bucket,omitempty"`
 	Kind         CacheContentKind `json:"kind,omitempty"`
 	CheckpointID string           `json:"checkpoint_id,omitempty"`
 	Accelerator  string           `json:"accelerator,omitempty"`
